Move styling to seperate file & remove unused classes
